Add userLocation variable
Add userLocation variable mapped to org.eclipse.equinox.p2.cache; this is needed
in multi-user installs where the installFolder variable doesn't change to point
to the location under the users home folder.

The use case here is to be able to use
org.eclipse.equinox.p2.touchpoint.natives.remove to delete files added when
installing a feature as a non-admin user.

In this case, the files are not added under ${installLocation} and can not be
removed via p2.